Switched database to mySQL from Docker image

+++ app/routes.py
@@ -1,19 +1,14 @@
 import json
 from flask import jsonify, request
-from flask_bcrypt import Bcrypt
 from app.models import User, Product
-from app import app, db
-
-bcrypt = Bcrypt(app)
+from app import app, db, bcrypt
+from flask_login import login_user, current_user, logout_user, login_required
 
 @app.route("/", methods=['GET'])
+#@login_required
 def home():
-    
-    # price = request.args.get('price', 'All')
-    # event = request.args.get('event', 'All')
-
-    all_products =  Product.query.all()
-
+    #all_products =  Product.query.filter_by(user_id=current_user.id).all()
+    all_products = Product.query.all()
     product_array = []
 
     for product in all_products:
@@ -32,6 +27,7 @@ def home():
     return jsonify(product_array)
 
 @app.route('/add_item', methods=['POST'])
+@login_required
 def add_item():
     req_data = request.get_json()
 
@@ -41,9 +37,9 @@ def add_item():
     quantity = req_data.get('quantity')
     price = req_data.get('price')
     picture = req_data.get('pic_location')
-    user_id = req_data.get('user_id')
+    user_id = current_user.id
 
-    temp_product = Product(item_name=name, school_event=event, storage_location=location, quantity=quantity, price=price, pic_location=picture)
+    temp_product = Product(item_name=name, school_event=event, storage_location=location, quantity=quantity, price=price, pic_location=picture, user_id=user_id)
 
     db.session.add(temp_product)
     db.session.commit()
@@ -52,23 +48,26 @@ def add_item():
     return jsonify({"status" : "success"}), 200
 
 @app.route('/delete_item', methods=['DELETE'])
+@login_required
 def delete_item():
     id = request.args.get('id')
     if id != None:
         item_to_delete = Product.query.get(id)
-        db.session.delete(item_to_delete)
-        db.session.commit()
-        return jsonify({"status": "item deleted"}), 200
+        if item_to_delete.user_id==current_user.id:
+            db.session.delete(item_to_delete)
+            db.session.commit()
+            return jsonify({"status": "item deleted"}), 200
     else: 
         return jsonify({"status" : "none deleted"})
 
 @app.route('/edit_item', methods=['PUT'])
+@login_required
 def edit_item():
     req_data = request.get_json()
     id = req_data.get('id')
     if id != None:
         item_to_edit = Product.query.get(id)
-        if item_to_edit!=None:
+        if item_to_edit!=None and item_to_edit.user_id==current_user.id:
             name = req_data.get('name')
             event = req_data.get('event')
             location = req_data.get('location')
@@ -92,12 +91,48 @@ def register():
     username = req_data.get('username')
     password = req_data.get('password')
     hashed_pass = bcrypt.generate_password_hash(password).decode('utf-8')
-    user = User(username=username, password=hashed_pass)
-    db.session.add(user)
-    db.session.commit()
-    return jsonify({"status": "user created"}), 200
+    user = User.query.filter_by(username=username).first()
+    status = "Error: that username is already taken"
+    if user==None:
+        user = User(username=username, password=hashed_pass)
+        db.session.add(user)
+        db.session.commit()
+        status = "success"
+    return status
 
-@app.route('/get_users', methods=['GET'])
+@app.route('/login', methods=['POST'])
+def login():
+    req_data = request.get_json()
+    username = req_data.get('username')
+    password = req_data.get('password')
+
+    status = "Error: incorrect username or password"
+    user = User.query.filter_by(username=username).first()
+    if user and bcrypt.check_password_hash(user.password, password):
+        login_user(user, remember=True)
+        if current_user.is_authenticated:
+            status = "success"
+
+    return status
+
+@app.route('/logout')
+@login_required
+def logout():
+    logout_user()
+    return jsonify({"status": "logged out"})
+
+
+@app.route('/this_user')
+def this_user():
+    if current_user.is_authenticated:
+        user = {
+            "id" : current_user.id,
+            "username" : current_user.username
+        }
+        return jsonify(user)
+    return "No user logged in"
+
+@app.route('/get_users')
 def get_users():
     all_users =  User.query.all()
     user_array = []
@@ -111,8 +146,3 @@ def get_users():
         user_array.append(temp_dict)
         
     return jsonify(user_array)
-
-@app.route('/get_user', methods=['GET'])
-def get_user():
-    req_data = request.get_json()
-    username = req_data.get('username')
